// src/types/vehicles.d.ts
export interface VehiclesProps {
    vei_codigo: number;
    vei_modelo: string;
    vei_placa: string;
    vei_ano: number;
    vei_status: string;
    vei_odometro: number;
    vei_prefixo: number;
    vei_imagem: string;
    emp_codigo: number;
    fro_codigo: number;
    fro_nome: string;
    vei_updated_at: string
    emp_nome: string;
}